+++ src/sprite.rs
@@ -7,8 +7,8 @@ use crate::movement::GMMovementT;
 
 pub struct GMSprite {
     texture: Rc<GMTexture>,
-    pub x: i32,
-    pub y: i32,
+    pub x: f32,
+    pub y: f32,
     pub active: bool,
     animation: Box<dyn GMAnimationT>,
     movement: Box<dyn GMMovementT>,
