+++ src/texture.rs
@@ -13,7 +13,7 @@ pub struct GMTexture {
 }
 
 impl GMTexture {
-    pub fn draw(&self, dx: i32, dy: i32, index: u32) {
+    pub fn draw(&self, dx: f32, dy: f32, index: u32) {
         let yi = index / self.cols;
         let xi = index - (yi * self.cols);
 
@@ -21,12 +21,12 @@ impl GMTexture {
         let sy = (yi * self.unit_height) as i32;
 
         let src_rect = Rect::new(sx, sy, self.unit_width, self.unit_height);
-        let dst_rect = Rect::new(dx, dy, self.unit_width, self.unit_height);
+        let dst_rect = Rect::new(dx as i32, dy as i32, self.unit_width, self.unit_height);
 
 
     }
 
-    pub fn get_unit_dimension(&self) -> (u32, u32) {
-        (self.unit_width, self.unit_height)
+    pub fn get_unit_dimension(&self) -> (f32, f32) {
+        (self.unit_width as f32, self.unit_height as f32)
     }
 }
